Extract nanoc-spec gem
