+++ nanoc-spec/.rspec
@@ -1,0 +1,3 @@
+-r ./spec/spec_helper.rb
+--format Fuubar
+--color
