+++ nanoc-spec/NEWS.md
@@ -1,0 +1,5 @@
+# nanoc-spec news
+
+## 0.0.1 (2019-11-16)
+
+Initial release.
