+++ nanoc-spec/README.md
@@ -1,0 +1,17 @@
+# nanoc-spec
+
+This provides `Nanoc::Spec`, which contains a bunch of functionality that facilitates writing tests for [Nanoc](https://nanoc.ws).
+
+## Installation
+
+Add `nanoc-spec` to the `nanoc` group of your Gemfile:
+
+```ruby
+group :nanoc do
+  gem 'nanoc-spec'
+end
+```
+
+## Usage
+
+Don’t — for now.
