+++ nanoc-spec/lib/nanoc-spec.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+require 'nanoc/spec'
