+++ nanoc-spec/lib/nanoc/spec/version.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Nanoc
+  module Spec
+    VERSION = '0.0.1'
+  end
+end
