+++ nanoc-spec/spec/spec_helper.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+require_relative '../../common/spec/spec_helper_head'
+
+require 'nanoc/spec'
+
+require_relative '../../common/spec/spec_helper_foot'
